pyproject.toml
[tool.poetry]
name = "deckster"
version = "0.1.0"
description = "Generate formatted PPT from JSON."
authors = ["jaredkirby <jpeterkirby@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-pptx = "^0.6.23"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
